src/main/webapp/app/entities/danh-muc-huyen/delete/danh-muc-huyen-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DanhMucHuyenService } from '../service/danh-muc-huyen.service';

import { DanhMucHuyenDeleteDialogComponent } from './danh-muc-huyen-delete-dialog.component';

describe('DanhMucHuyen Management Delete Component', () => {
  let comp: DanhMucHuyenDeleteDialogComponent;
  let fixture: ComponentFixture<DanhMucHuyenDeleteDialogComponent>;
  let service: DanhMucHuyenService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DanhMucHuyenDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DanhMucHuyenDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DanhMucHuyenDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DanhMucHuyenService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
